Implementing an abstract class for the OptimizerGraphViewModel, DRY
