--- src/HeatManager/ViewModels/OptimizerGraphs/IOptimizerCostsGraph.cs
@@ -1,5 +1,0 @@
-namespace HeatManager.ViewModels.OptimizerGraphs;
-
-interface IOptimizerCostsGraph
-{
-}
